requirements.txt
flask
python-dotenv
beautifulsoup4
requests
langchain-google-genai